refactor battleCLI init and DOM
